# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OCC Operational Control Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /activity-logs:
    get:
      operationId: listActivityLogs
      tags: [activityLogs]
      summary: List activity logs
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: ptId
          in: query
          schema:
            type: integer
        - name: shiftId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Activity logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLogWithRelations"
    post:
      operationId: createActivityLog
      tags: [activityLogs]
      summary: Create activity log
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityLogRequest"
      responses:
        "201":
          description: Activity log created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityLog"
  /activity-logs/batch:
    post:
      operationId: createBatchActivityLogs
      tags: [activityLogs]
      summary: Create multiple activity logs at once
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchActivityLogRequest"
      responses:
        "201":
          description: Activity logs created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"
  /activity-logs/{id}:
    put:
      operationId: updateActivityLog
      tags: [activityLogs]
      summary: Update activity log (time-limited)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActivityLogRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityLog"
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      summary: Get announcement by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Announcement details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    put:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    ActivityLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        activityTypeId:
          type: integer
        quantity:
          type: integer
        note:
          type: string
        ptId:
          type: integer
        branchId:
          type: integer
        shiftId:
          type: integer
        points:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, activityTypeId, quantity, points]
    ActivityLogWithRelations:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        activityTypeId:
          type: integer
        activityTypeName:
          type: string
        quantity:
          type: integer
        note:
          type: string
        ptId:
          type: integer
        branchId:
          type: integer
        shiftId:
          type: integer
        points:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, activityTypeId, quantity, points]
    CreateActivityLogRequest:
      type: object
      properties:
        activityTypeId:
          type: integer
        quantity:
          type: integer
        note:
          type: string
      required: [activityTypeId, quantity]
    BatchActivityLogRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateActivityLogRequest"
      required: [items]
    UpdateActivityLogRequest:
      type: object
      properties:
        quantity:
          type: integer
        note:
          type: string
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        targetScope:
          type: string
        ptId:
          type: integer
        branchId:
          type: integer
        shiftId:
          type: integer
        roleId:
          type: integer
        createdBy:
          type: integer
        creatorName:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        priority:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, content, priority]
    CreateAnnouncementRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        targetScope:
          type: string
        ptId:
          type: integer
        branchId:
          type: integer
        shiftId:
          type: integer
        roleId:
          type: integer
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        priority:
          type: string
      required: [title, content]
...